Fix segfault when a unit file changes
A previous removed the need for daemon-reload, because we just
actually reread units which need to be reread instead of passively
aggressively telling the user they need to run daemon-reload:

dcbf64216c6d57b8c73f8eaabfa628e602e5cac6

However, this was causing a segfault when a unit file changed because
argv[] was not set in this case.

Setting argv[] to "Reload" fixes this crash.